Refactoring Timestamp.cpp It contains more responsabilities than desired. Splitted the content in three 	Date.cpp, that holds the Date class 	Timestamp.cpp, that holds the Timestamp class 	TimeFunctions.cpp, that contains the definition for Vireo Functions
