+++ Vireo_VS/VireoCommandLine.vcxproj
@@ -16,6 +16,7 @@
     <ClCompile Include="..\source\core\Assert.cpp" />
     <ClCompile Include="..\source\core\CEntryPoints.cpp" />
     <ClCompile Include="..\source\core\DataQueue.cpp" />
+    <ClCompile Include="..\source\core\Date.cpp" />
     <ClCompile Include="..\source\core\EventLog.cpp" />
     <ClCompile Include="..\source\core\ExecutionContext.cpp" />
     <ClCompile Include="..\source\core\GenericFunctions.cpp" />
@@ -29,6 +30,7 @@
     <ClCompile Include="..\source\core\TDCodecLVFlat.cpp" />
     <ClCompile Include="..\source\core\TDCodecVia.cpp" />
     <ClCompile Include="..\source\core\Thread.cpp" />
+    <ClCompile Include="..\source\core\TimeFunctions.cpp" />
     <ClCompile Include="..\source\core\Timestamp.cpp" />
     <ClCompile Include="..\source\core\TypeAndDataManager.cpp" />
     <ClCompile Include="..\source\core\TypeAndDataReflection.cpp" />
@@ -42,8 +44,6 @@
     <ClCompile Include="..\source\io\WebSocketClient.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="..\source\core\vireo.postamble.js" />
-    <None Include="..\source\core\vireo.preamble.js" />
     <None Include="..\source\include\ConversionTable.def" />
     <None Include="..\source\io\library_canvas2d.js" />
     <None Include="..\source\io\library_httpClient.js" />
@@ -53,9 +53,11 @@
     <ClInclude Include="..\source\include\BuildConfig.h" />
     <ClInclude Include="..\source\include\CEntryPoints.h" />
     <ClInclude Include="..\source\include\DataTypes.h" />
+    <ClInclude Include="..\source\include\Date.h" />
     <ClInclude Include="..\source\include\EventLog.h" />
     <ClInclude Include="..\source\include\ExecutionContext.h" />
     <ClInclude Include="..\source\include\Instruction.h" />
+    <ClInclude Include="..\source\include\LVDateTimeRecord.h" />
     <ClInclude Include="..\source\include\Platform.h" />
     <ClInclude Include="..\source\include\StringUtilities.h" />
     <ClInclude Include="..\source\include\Synchronization.h" />
