+++ source/include/Timestamp.h
@@ -8,7 +8,7 @@ SDG
 */
 
 /*! \file
-    \brief Tools to working with relative, absolute and tic-count times.
+    \brief Timestamp class definition.
  */
 
 #ifndef Timestamp_h
@@ -17,22 +17,22 @@ SDG
 #include "DataTypes.h"
 
 namespace Vireo {
-
 //------------------------------------------------------------
 //! A 64.64 bit fixed point use to represent seconds since Jan 1, 1904 GMT
-class Timestamp
-{
- private:
+class Timestamp {
+private:
     Int64 _integer;
     UInt64 _fraction;
 
- public:
+public:
     Timestamp()
         { _integer = 0; _fraction = 0;}
     Timestamp(Int64 integer, UInt64 fraction)
         { _integer = integer; _fraction = fraction;}
     Timestamp(Double seconds);
     Timestamp(Double fracSecs, Int32 sec, Int32 min, Int32 hour, Int32 day, Int32 month, Int32 year);
+
+    static void GetCurrentTimestamp(Timestamp *t);
 
     Int64 Integer() const { return _integer; };
     UInt64 Fraction() const { return _fraction; };
@@ -67,53 +67,5 @@ class Timestamp
     }
     Double ToDouble (void) const;
 };
-
-class Date
-{
- private:
-    Double _fractionalSecond;
-    Int32 _second;
-    Int32 _minute;
-    Int32 _hour;
-    Int32 _day;
-    Int32 _month;
-    Int32 _year;
-    Int32 _weekday;
-    Int64 _secondsOfYear;
-    Int32 _firstWeekDay;
-    Int32 _timeZoneOffset;
-    ConstCStr _timeZoneString;
-    Int32 _DTS;
-    static Int32 _SystemLocaletimeZone;
- public:
-        Date(Timestamp timestamp, Int32 timeZone);
-        static Int32 getLocaletimeZone();
-        Int32 Year() const { return _year; };
-        Int32 Month() const { return _month; };
-        Int32 Day() const { return _day; };
-        Int32 Hour() const { return _hour; };
-        Int32 Minute() const { return _minute; };
-        Int32 Second() const { return _second; };
-        Int32 WeekDay() const { return _weekday; };
-        Int32 FirstWeekDay() const { return _firstWeekDay; };
-        Int64 SecondsOfYear() const {return _secondsOfYear;};
-        Int32 TimeZoneOffset() const {return _timeZoneOffset;};
-        ConstCStr TimeZoneString() const { return _timeZoneString; };
-        Int32 DTS() const {return _DTS;}
-        Int32 isDTS();
-        Double FractionSecond() const {return  _fractionalSecond;};
-        // return a new date object at other time zone
-        Date AtTimeZone(Int32 timeZone);
-};
-
-struct LVDateTimeRec {
-    Double fractional_secs;
-    Int32 second, minute, hour;
-    Int32 day_of_month, month, year;
-    Int32 day_of_week, day_of_year, dst;
-
-    LVDateTimeRec() : fractional_secs(0), second(0), minute(0), hour(0), day_of_month(0), month(0), year(0),
-        day_of_week(0), day_of_year(0), dst(0) {}
-};
-}  // namespace Vireo
+}
 #endif // Timestamp_h
